Implement Good Bot and Bad Bot Responses
When Suddenly, Geese! receives a 'good bot' it will now honk seductively. Likewise, when it receives a bad bot, it will hiss. Additionally, the goose detector regex has been improved to ignore any instances of geese that are not on a word boundary or are part of a hyphenated word - that is to say, 'goose-step' and 'mongoose' will be ignored. Lastly code has been cleaned up and readability improved.